adding webpack to the gulp task for building the lib
We need webpack because we want the lib to be able to be used in different ways such as loading it via a script tag or using a ES2015 import. Axios is now a dev dependecy because it's needed for the develpment of the lib and it's added as a peer dependency because we want the consumer of the lib to already have axios installed. This will prevent packaging two versions of axios if someone had already added axios to their project.